feat(docker-tests): add ETH test split with docker-tests-eth feature
- Add docker-tests-eth feature flag to Cargo.toml
- Add docker-tests-eth CI job (60 min timeout, evm profile)
- Extract shared ETH helpers to helpers/eth.rs module
- Update imports in dependent test modules
- Feature-gate eth_docker_tests module
- Update docs/DOCKER_TESTS.md with ETH mapping
- Add root CLAUDE.md reference note to all crate AGENTS.md files
