+++ mm2src/mm2_main/Cargo.toml
@@ -21,6 +21,7 @@ run-docker-tests = ["coins/run-docker-tests"]
 # Split docker test features - each enables a specific test group
 docker-tests-slp = ["run-docker-tests"]
 docker-tests-sia = ["run-docker-tests"]
+docker-tests-eth = ["run-docker-tests"]
 default = []
 trezor-udp = ["crypto/trezor-udp"] # use for tests to connect to trezor emulator over udp
 run-device-tests = []
