+++ mm2src/mm2_main/tests/docker_tests/docker_tests_inner.rs
@@ -1,7 +1,7 @@
 use crate::docker_tests::docker_tests_common::{
     generate_utxo_coin_with_privkey, trade_base_rel, GETH_RPC_URL, MM_CTX, SET_BURN_PUBKEY_TO_ALICE,
 };
-use crate::docker_tests::eth_docker_tests::{
+use crate::docker_tests::helpers::eth::{
     erc20_coin_with_random_privkey, erc20_contract_checksum, fill_eth_erc20_with_private_key, swap_contract,
 };
 use crate::integration_tests_common::*;
