+++ mm2src/mm2_main/tests/docker_tests/helpers/mod.rs
@@ -1,0 +1,4 @@
+//! Shared helper functions for docker tests.
+//! These helpers are available to all test modules regardless of feature flags.
+
+pub mod eth;
